refactor: moved reducer
