+++ src/contexts/FavoritesContext.tsx
@@ -1,5 +1,6 @@
 import { createContext, useContext, useEffect, useReducer } from "react";
 import type { Job } from "../models/Job";
+import { favoritReducer } from "../reducers/favoritReducer";
 
 export const FavoritContext = createContext<fav | undefined>(undefined);
 
@@ -10,29 +11,8 @@ type fav = {
   isFavorite: (id: string | number) => boolean;
 };
 
-type Action =
-  |{ type: "ADD"; job: Job }
-  |{ type: "REMOVE"; id: string | number };
-
-function reducer(state: Record<string, Job>, action: Action): Record<string, Job> {
-  switch (action.type) {
-  case "ADD": {
-      const favKey = String((action.job as any).id);
-      return state[favKey] ? state : { ...state, [favKey]: action.job };
-    }
-  case "REMOVE": {
-      const favKey = String(action.id);
-  if (!state[favKey]) return state;
-      const { [favKey]: _, ...rest } = state;
-  return rest;
-    }
-    default:
-      return state;
-  }
-}
-
 export function FavoritProvider({ children }: { children: React.ReactNode }) {
-  const [favoritesMap, dispatch] = useReducer(reducer, {}, () => {
+  const [favoritesMap, dispatch] = useReducer(favoritReducer, {}, () => {
     try {
       const raw = localStorage.getItem("favorites");
       return raw ? (JSON.parse(raw) as Record<string, Job>) : {};
@@ -47,11 +27,9 @@ export function FavoritProvider({ children }: { children: React.ReactNode }) {
     } catch {}
   }, [favoritesMap]);
 
-  const addFavorite = (job: Job) => 
-        dispatch({ type: "ADD", job });
+  const addFavorite = (job: Job) =>dispatch({ type: "ADD", job });
 
-  const removeFavorite = (id: string | number) =>
-  dispatch({ type: "REMOVE", id });
+  const removeFavorite = (id: string | number) => dispatch({ type: "REMOVE", id });
 
   const isFavorite = (id: string | number) => {
     const favKey = String(id);
